Extract Instabug to its own Module
